fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
transformers==4.35.0
python-docx==0.8.11
python-multipart==0.0.6
accelerate==0.23.0